requirements.txt
numpy==1.24.3
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
chromadb==0.4.15
pypdf2==3.0.1
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic>=2.7.0
pydantic-settings>=2.2.0
pytest==7.4.3
httpx==0.25.2 
